+++ app/projector/src/commonMain/kotlin/org/hnau/commons/app/projector/fractal/utils/BaseWithDecayContrastExt.kt
@@ -6,7 +6,7 @@ import org.hnau.commons.app.model.theme.color.Contrast
 fun BaseWithDecay.Companion.contrast(
     initial: Contrast,
     decay: Double,
-): BaseWithDecay<Contrast> = BaseWithDecay.int(
+): BaseWithDecay<Contrast> = BaseWithDecay.double(
     initial = initial.contrast,
     decay = decay,
     baseline = Contrast.min.contrast
@@ -15,7 +15,7 @@ fun BaseWithDecay.Companion.contrast(
 
 
 private val containerContrast: BaseWithDecay<Contrast> = BaseWithDecay.contrast(
-    initial = Contrast(3),
+    initial = Contrast(3.0),
     decay = 0.8,
 )
 
@@ -24,11 +24,11 @@ val Contrast.Companion.container: BaseWithDecay<Contrast>
 
 private val contentBySaturationContrast: SaturationValues<BaseWithDecay<Contrast>> = SaturationValues(
     active = BaseWithDecay.contrast(
-        initial = Contrast(9),
+        initial = Contrast(9.0),
         decay = 0.9,
     ),
     neutral = BaseWithDecay.contrast(
-        initial = Contrast(7),
+        initial = Contrast(7.0),
         decay = 0.9,
     ),
 )
@@ -39,7 +39,7 @@ val Contrast.Companion.contentBySaturation: SaturationValues<BaseWithDecay<Contr
     get() = contentBySaturationContrast
 
 private val contentLowContrast: BaseWithDecay<Contrast> = BaseWithDecay.contrast(
-    initial = Contrast(7),
+    initial = Contrast(7.0),
     decay = 0.9,
 )
